# backend/openapi.yml
openapi: 3.0.3
info:
  title: Lojinha do Zé - API
  description: |
    API REST para o sistema de gestão da Lojinha do Zé (e-commerce de produtos naturais e artesanais).

    ## Autenticação

    A API utiliza autenticação baseada em cookies de sessão. Após o login, o cookie `session_id` deve ser incluído nas requisições subsequentes.

    ### Fluxo de Autenticação
    1. Faça login em `/api/auth/login` para receber um `session_id` cookie e `csrfToken`
    2. Inclua o cookie `session_id` automaticamente (via browser)
    3. Inclua o header `X-CSRF-Token: <csrfToken>` em requisições POST/PUT/PATCH/DELETE

    ## Rate Limiting

    | Endpoint | Limite | Janela |
    |----------|--------|--------|
    | Login | 5 tentativas | 15 minutos |
    | Pedidos | 10 pedidos | 1 hora |
    | API Geral | 60 requisições | 1 minuto |

    ## Códigos de Status

    - `200` - Sucesso
    - `201` - Criado
    - `400` - Requisição Inválida
    - `401` - Não Autenticado
    - `403` - Proibido (sem permissão)
    - `404` - Não Encontrado
    - `409` - Conflito (duplicado)
    - `429` - Rate Limit Excedido
    - `500` - Erro Interno
  version: 1.0.0
  contact:
    name: SHURIKA6
    url: https://github.com/SHURIKA6/Lojinha-do-Ze
servers:
  - url: http://localhost:8787/api
    description: Desenvolvimento Local
  - url: https://lojinha-do-ze-backend.your-subdomain.workers.dev/api
    description: Produção (Cloudflare Workers)
...
paths:
...
  /catalog:
    get:
      tags: [Catálogo]
      summary: Lista produtos do catálogo
      description: Retorna produtos agrupados por categoria, com filtros opcionais
      operationId: getCatalog
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          schema:
            type: string
          description: Busca por nome do produto
        - name: category
          in: query
          schema:
            type: string
          description: Filtra por categoria
      responses:
        '200':
          description: Catálogo de produtos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
  /catalog/orders:
    post:
      tags: [Catálogo]
      summary: Cria novo pedido
      description: |
        Cria um pedido para cliente autenticado ou convidado.

        - Se autenticado como `customer`, o `customer_phone` deve bater com o usuário
        - Se `delivery_type = 'entrega'`, o campo `address` é obrigatório
        - Estoque é reservado no momento da criação
      operationId: createOrder
      security: [{ SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Pedido criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: '#/components/schemas/Order'
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
  /orders:
    get:
      tags: [Pedidos]
      summary: Lista pedidos
      description: |
        - Admin: vê todos os pedidos
        - Customer: vê apenas seus pedidos
      operationId: listOrders
      security: [{ SessionAuth: [] }]
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
...
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: cookie
      name: session_id
      description: Cookie de sessão após login
    AdminOnly:
      type: apiKey
      in: header
      name: X-Admin-Check
      description: Requer papel de administrador (verificado via sessão)
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true

    # Auth
...
    CatalogResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              products:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        availableCategories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              count:
                type: integer
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    Product:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        photo:
          type: string
        category:
          type: string
        sale_price:
          type: string
        quantity:
          type: integer
        is_active:
          type: boolean

    # Pedidos
    Order:
      type: object
      properties:
        id:
          type: integer
        customer_id:
          type: string
          nullable: true
        customer_name:
          type: string
        customer_phone:
          type: string
        items:
          type: string
          description: JSON stringified de OrderItem[]
        subtotal:
          type: number
        delivery_fee:
          type: number
        total:
          type: number
        status:
          $ref: '#/components/schemas/OrderStatus'
        delivery_type:
          type: string
          enum: [retirada, entrega]
        address:
          type: string
        payment_method:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OrderCreate:
      type: object
      required: [customer_name, customer_phone, items, delivery_type, payment_method]
      properties:
        customer_name:
          type: string
          minLength: 2
        customer_phone:
          type: string
          minLength: 8
        notes:
          type: string
        delivery_type:
          type: string
          enum: [retirada, entrega]
        address:
          type: string
        payment_method:
          type: string
          enum: [pix, dinheiro, cartao]
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: integer
              quantity:
                type: integer
                minimum: 1
    OrderStatus:
      type: string
      enum: [novo, recebido, em_preparo, saiu_entrega, entregue, concluido, cancelado]

    # Dashboard
...